fix(next): 敵対的レビューの指摘を潰す — 軸の破れ4件と柵の空振り
利用者指示の「軸のドリフト確認」で独立レビューを回し、実害を潰した。

## 軸が破れていた

- **時刻の正本が2本になっていた**: engine が `t.num() as f64 / t.den() as f64 *
  fps.num() as f64 / ...` で frame を出しており、`motolii-core` の
  「f64×fps の独自丸めは禁止」に真っ向から反していた。preview も export も同じ
  関数を通るので **M15 は保たれたまま両方同じようにずれる**。試験が 30/1 でしか
  回っていないので NTSC を入れるまで気付けなかった。正準口 `try_to_frame_floor` へ
- **柵が構造でなく宣言だった**: 旧 TM-4 は禁止パターンを5個の文字列で探しており、
  上の式はどれにも一致せず素通りした。**リセット裁定が旧 workspace を断罪したのと
  同じ型の失敗が当日に再発**。パターン列挙をやめ「fps と f64 が同じ式に出ること」
  自体を禁じる柵へ(裁定32)
- **export が第二の合成器を建てられた**: `CompSpec` を取るためだけに
  `motolii-compositor` を引いていた。型を core へ出して依存を切り、
  依存グラフを見る柵を置いた(裁定33)
- **`motolii-store` の marker が誤り**: `wraps:` を名乗る crate に `fingerprint.rs` と
  `resolve` という `owns:` の中身が入っていた。crate の根しか見ない check.sh が
  空振りしていた(裁定34)

## 実害

- 素材の外の時刻で**フレーム全体が落ちていた**。コメントは「描かない」と書いてあるのに
  実装は上限を見ていなかった。`nb_frames` を見てその layer だけ落とす(裁定38)
- `PropertyId::new("meta")` が layer の素材と重ね順を**上書きできた**(裁定35)
- `order: i32` → `DepthOffset: i16` の無言切り捨て。32768 以上で符号反転(裁定36)
- 読み口の4箇所で「読めない」を「無い」と同義にしていた。リセット裁定が旧 iced を
  「拒否を let _ = で捨てていた」と名指しした同じ形(裁定37)
- フレームキャッシュが FIFO で、窓が layer 数より小さいと毎フレーム全 evict になる
  構造だった。LRU にし、窓の根拠を「同時に描ける media layer 数」に

## 台帳の嘘を正した

- GOALS M15/D1「byte 一致」→ 実際は Y の最大差 ≤ 8(h264 が YUV420 を通る)
- GOALS M14「済」→ **未**。選択・playhead・fps・解像度・尺が store に1つも無い
- GOALS M10「済」→ **部分**。gesture を1 undo へ畳む口が無い
- 裁定27 の「別 device でも一致」→ iced の device は1度も通していないので過大

自分でも1件やらかした: `try_from_frame` が既に core にあるのに同名を足して重複定義に
なった。**まさに潰したい失敗型**なので、既存の doc に「口が在るのに使われないのが
実際の失敗の形」と書き残した。

182 passed / 4 ignored。

+++ next/probes/r1-frame-throughput/tests/r1.rs
@@ -46,7 +46,7 @@ fn document_scaled(layers: u64, source_scale: u32) -> Document {
                     width: sw,
                     height: sh,
                 },
-                order: i as i32,
+                order: i as i16,
             },
         })
         .unwrap();
@@ -68,7 +68,7 @@ fn document_with(layers: u64) -> Document {
                     width: FHD[0],
                     height: FHD[1],
                 },
-                order: i as i32,
+                order: i as i16,
             },
         })
         .unwrap();
